feat: add more insight api

+++ database/migrations/2025_12_12_151458_create_reports_table.php
@@ -15,29 +15,29 @@ return new class extends Migration {
             $table->tinyInteger('quarter');
             $table->smallInteger('year');
             $table->integer('member_growth_in');
-            $table->integer('member_growth_in_percentage');
+            $table->float('member_growth_in_percentage');
             $table->integer('member_growth_out');
-            $table->integer('member_growth_out_percentage');
+            $table->float('member_growth_out_percentage');
             $table->integer('group_member_total');
-            $table->integer('group_member_total_percentage');
-            $table->integer('administrative_compliance_percentage');
-            $table->integer('deposit_compliance_percentage');
-            $table->integer('attendance_percentage');
-            $table->integer('organization_final_score_percentage');
+            $table->float('group_member_total_percentage');
+            $table->float('administrative_compliance_percentage');
+            $table->float('deposit_compliance_percentage');
+            $table->float('attendance_percentage');
+            $table->float('organization_final_score_percentage');
             $table->integer('loan_participation_pb');
             $table->integer('loan_participation_bbm');
             $table->integer('loan_participation_store');
             $table->integer('cash_participation');
-            $table->integer('cash_participation_percentage');
+            $table->float('cash_participation_percentage');
             $table->integer('savings_participation');
-            $table->integer('savings_participation_percentage');
-            $table->integer('meeting_deposit_percentage');
+            $table->float('savings_participation_percentage');
+            $table->float('meeting_deposit_percentage');
             $table->integer('loan_balance_pb');
             $table->integer('loan_balance_bbm');
             $table->integer('loan_balance_store');
-            $table->integer('receivable_score');
-            $table->integer('financial_final_score_percentage');
-            $table->integer('combined_final_score_percentage');
+            $table->float('receivable_score');
+            $table->float('financial_final_score_percentage');
+            $table->float('combined_final_score_percentage');
             $table->enum('criteria', [
                 'sangat_baik',
                 'baik',
